Enabling parsing of reference tests #9
This now enables a minimal "build" of the reference tests using the
WhileyFile library.  Specifically, all it does at this stage is attempt
to parse each Whiley file in each snapshot of a test file.  Of course,
   this will fail when a parsing error is expected.  However, it fails
   also in most cases now simply because the parser is not powerful
   enough.

There are a few ways the API used within the reftests driver could be
improved.  Firstly, allowing a WhileyTestFile to have an iterator would
be nice.  Likewise, removing the content of SourceFile and merging that
into WhileyFile.  The latter requires some API changes to WhileyFile,
     but at the same time paves the way for incremental computation.