Partial support for support mapping
This adds partial support for source mapping, though leaves the more
complicated cases still to do.  One of the problems here is that there's
nothing to actually use source mapping yet, so its hard to check it.

Another issue is that its not clear to me how to integrate SourceMap
into WhileyFile.  I think that's important, so it remains to do.